+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
 - Handle when there's no changelog file in the repo ([#15](https://github.com/SmartBear/changelog-bot/issues/15))
 - Run a full scan on installation if there is already a changelog in the repository ([#16](https://github.com/SmartBear/changelog-bot/issues/16))
+
+### Fixed
+- More robust changelog anchor generation ([#18](https://github.com/SmartBear/changelog-bot/issues/18))
 
 ## [0.0.1] - 2021-10-06
 ### Fixed